- Fixed issue with loading some JSON files, seemed to be caused by using C style file reading, changed to C++ style. - Fixed some casting warnings. - Added 'lite' version of the transform class, which is designed to work in just 2D. It also only takes up 12 bytes of memory instead of 128 used by the full 3D transform class which was a waste.

+++ gfx/gfx_asset.cpp
@@ -265,7 +265,7 @@ namespace gfx {
 				right = 0.0f;
 				if (mTexture)
 				{
-					right = mTexture->getWidth();
+					right = static_cast<float>(mTexture->getWidth());
 				}
 			}
 			if (top < 0.0f || bottom < 0.0f)
@@ -274,7 +274,7 @@ namespace gfx {
 				bottom = 0.0f;
 				if (mTexture)
 				{
-					bottom = mTexture->getHeight();
+					bottom = static_cast<float>(mTexture->getHeight());
 				}
 			}
